# docs/openapi.json
{
  "components": {
    "schemas": {
      "GPU": {
        "properties": {
          "Hostname": {
            "type": "string"
          },
          "device": {
            "type": "string"
          },
          "gpu_id": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "modelName": {
            "type": "string"
          },
          "uuid": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "GPUListResponse": {
        "properties": {
          "items": {
            "items": {
              "properties": {
                "Hostname": {
                  "type": "string"
                },
                "device": {
                  "type": "string"
                },
                "gpu_id": {
                  "type": "string"
                },
                "id": {
                  "type": "string"
                },
                "modelName": {
                  "type": "string"
                },
                "uuid": {
                  "type": "string"
                }
              },
              "type": "object"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "GPUTelemetryResponse": {
        "properties": {
          "gpu_id": {
            "type": "string"
          },
          "items": {
            "items": {
              "properties": {
                "Hostname": {
                  "type": "string"
                },
                "container": {
                  "type": "string"
                },
                "device": {
                  "type": "string"
                },
                "gpu_id": {
                  "type": "string"
                },
                "labels_raw": {
                  "type": "string"
                },
                "metric_name": {
                  "type": "string"
                },
                "modelName": {
                  "type": "string"
                },
                "namespace": {
                  "type": "string"
                },
                "pod": {
                  "type": "string"
                },
                "timestamp": {
                  "format": "date-time",
                  "type": "string"
                },
                "uuid": {
                  "type": "string"
                },
                "value": {
                  "type": "number"
                }
              },
              "type": "object"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "TelemetryQueryResponse": {
        "properties": {
          "items": {
            "items": {
              "properties": {
                "Hostname": {
                  "type": "string"
                },
                "container": {
                  "type": "string"
                },
                "device": {
                  "type": "string"
                },
                "gpu_id": {
                  "type": "string"
                },
                "labels_raw": {
                  "type": "string"
                },
                "metric_name": {
                  "type": "string"
                },
                "modelName": {
                  "type": "string"
                },
                "namespace": {
                  "type": "string"
                },
                "pod": {
                  "type": "string"
                },
                "timestamp": {
                  "format": "date-time",
                  "type": "string"
                },
                "uuid": {
                  "type": "string"
                },
                "value": {
                  "type": "number"
                }
              },
              "type": "object"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "TelemetrySample": {
        "properties": {
          "Hostname": {
            "type": "string"
          },
          "container": {
            "type": "string"
          },
          "device": {
            "type": "string"
          },
          "gpu_id": {
            "type": "string"
          },
          "labels_raw": {
            "type": "string"
          },
          "metric_name": {
            "type": "string"
          },
          "modelName": {
            "type": "string"
          },
          "namespace": {
            "type": "string"
          },
          "pod": {
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          },
          "uuid": {
            "type": "string"
          },
          "value": {
            "type": "number"
          }
        },
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Query processed GPU telemetry samples by time window and optional filters.",
    "title": "GPU Telemetry API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
    "/api/v1/gpus": {
      "get": {
        "operationId": "listGPUs",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GPUListResponse"
                }
              }
            },
            "description": "Available GPUs"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "List GPUs"
      }
    },
    "/api/v1/gpus/{id}/telemetry": {
      "get": {
        "operationId": "listGPUTelemetry",
        "parameters": [
          {
            "description": "GPU UUID.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Start of the query window in RFC3339 format. Required with end_time unless window is provided.",
            "in": "query",
            "name": "start_time",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "description": "End of the query window in RFC3339 format. Optional with window and defaults to now.",
            "in": "query",
            "name": "end_time",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "description": "Relative query window such as 5m, 15m, or 1h. Cannot be combined with start_time.",
            "in": "query",
            "name": "window",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional metric name filter.",
            "in": "query",
            "name": "metric_name",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional hostname filter.",
            "in": "query",
            "name": "hostname",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional GPU ID filter.",
            "in": "query",
            "name": "gpu_id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional device filter.",
            "in": "query",
            "name": "device",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional result limit.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GPUTelemetryResponse"
                }
              }
            },
            "description": "Telemetry query result"
          },
          "400": {
            "description": "Invalid query parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Query telemetry for a GPU"
      }
    },
    "/health": {
      "get": {
        "operationId": "getHealth",
        "responses": {
          "200": {
            "description": "API is healthy"
          }
        },
        "summary": "Health check"
      }
    },
    "/telemetry": {
      "get": {
        "deprecated": true,
        "operationId": "listTelemetry",
        "parameters": [
          {
            "description": "Start of the query window in RFC3339 format. Required with end unless window is provided.",
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "description": "End of the query window in RFC3339 format. Optional with window and defaults to now.",
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "description": "Relative query window such as 5m, 15m, or 1h. Cannot be combined with start.",
            "in": "query",
            "name": "window",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional metric name filter.",
            "in": "query",
            "name": "metric_name",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional GPU UUID filter.",
            "in": "query",
            "name": "uuid",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional hostname filter.",
            "in": "query",
            "name": "hostname",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional GPU ID filter.",
            "in": "query",
            "name": "gpu_id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional device filter.",
            "in": "query",
            "name": "device",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional result limit.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelemetryQueryResponse"
                }
              }
            },
            "description": "Telemetry query result"
          },
          "400": {
            "description": "Invalid query parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Query telemetry samples"
      }
    }
  }
}